+++ src/main/scala/edu/ie3/mobsim/model/EvType.scala
@@ -6,6 +6,7 @@
 
 package edu.ie3.mobsim.model
 
+import edu.ie3.datamodel.models.input.system.`type`.EvTypeInput
 import edu.ie3.mobsim.exceptions.InitializationException
 import edu.ie3.mobsim.io.probabilities.ProbabilityDensityFunction
 import edu.ie3.util.quantities.PowerSystemUnits
@@ -84,6 +85,18 @@ object EvType {
           dcPower
         )
       }
+  }
+
+  def apply(evTypeInput: EvTypeInput): EvType = {
+    new EvType(
+      "custom-model",
+      "custom-producer",
+      "custom-segment",
+      evTypeInput.geteStorage(),
+      evTypeInput.geteCons(),
+      evTypeInput.getsRated(),
+      evTypeInput.getsRated()
+    )
   }
 
   /** Determine the probabilities for each available ev model. It is based on