+++ src/test/scala/edu/ie3/mobsim/model/ElectricVehicleSpec.scala
@@ -15,6 +15,7 @@ import tech.units.indriya.quantity.Quantities
 import tech.units.indriya.unit.Units
 
 import java.time.ZonedDateTime
+import java.util.UUID
 import javax.measure.quantity.Energy
 import scala.collection.immutable.Queue
 
@@ -257,6 +258,54 @@ class ElectricVehicleSpec extends UnitSpec with TripSimulationTestData {
               givenSimulationStart,
               targetShare,
               givenModelPdf,
+              givenFirstDepartureMetaData
+            )
+
+            evs should have size targetAmount
+            evs.count(
+              _.chargingAtHomePossible
+            ) shouldBe expectedAmountOfHomeCharging
+        }
+      }
+
+      "work with EvInputs" in {
+        /* Create 200 home POI, where 100 are available for home charging */
+        val homePoisWithSizes = (Range(0, 100).map { cnt =>
+          givenHomePoi.copy(
+            id = s"home_poi_$cnt",
+            nearestChargingStations = Map(
+              givenChargingStation -> Quantities.getQuantity(1.0, Units.METRE)
+            )
+          )
+        } ++ Range(100, 200).map { cnt =>
+          givenHomePoi.copy(
+            id = s"home_poi_$cnt"
+          )
+        }).map(_ -> 1.0).toMap
+
+        val evInputs =
+          Range(0, 120).map(_ => evInput.copy().uuid(UUID.randomUUID()).build())
+
+        forAll(
+          Table(
+            ("targetAmount", "targetShare", "expectedAmountOfHomeCharging"),
+            (120, 1.0, 120),
+            (120, 0.1, 12),
+            (120, 0.12, 14)
+          )
+        ) {
+          case (
+                targetAmount: Int,
+                targetShare: Double,
+                expectedAmountOfHomeCharging: Int
+              ) =>
+            val evs = ElectricVehicle.createEvsFromEvInput(
+              evInputs,
+              homePoisWithSizes,
+              givenWorkPoiPdf,
+              Seq(givenChargingStation),
+              givenSimulationStart,
+              targetShare,
               givenFirstDepartureMetaData
             )